# swagger/tenants.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tenant api tests",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "localhost api server"
        }
    ],
    "components": {
        "schemas": {
            "apiResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "readOnly": true
                    },
                    "data": {
                        "type": "object",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "required": [
                    "success",
                    "message"
                ]
            },
            "TenantGetInfoRequest": {
                "type": "object",
                "properties": {
                    "tenant_id": {
                        "type": "string",
                        "readOnly": true,
                        "format": "uuid"
                    }
                },
                "required": [
                    "tenant_id"
                ],
                "example": {
                    "tenant_id": "f7dfed28-ebaa-4bc3-b556-b1787c74e554"
                }
            },
            "TenantGetInfoBySlugRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "name"
                ],
                "example": {
                    "name": "default"
                }
            },
            "TenantAddRequest": {
                "type": "object",
                "properties": {
                    "tenant_id": {
                        "type": "string",
                        "readOnly": true,
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "slug": {
                        "type": "string",
                        "readOnly": true
                    },
                    "description": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "tenant_id",
                    "name",
                    "slug",
                    "description"
                ],
                "example": {
                    "tenant_id": "f7dfed28-ebaa-4bc3-b556-b1787c74e554",
                    "name": "swagger test",
                    "slug": "swagger",
                    "description": "swagger test"
                }
            },
            "TenantUpdateRequest": {
                "type": "object",
                "properties": {
                    "tenant_id": {
                        "type": "string",
                        "readOnly": true,
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "slug": {
                        "type": "string",
                        "readOnly": true
                    },
                    "description": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "tenant_id",
                    "name",
                    "slug",
                    "description"
                ],
                "example": {
                    "tenant_id": "f7dfed28-ebaa-4bc3-b556-b1787c74e554",
                    "name": "swagger test update",
                    "slug": "swagger",
                    "description": "swagger test update"
                }
            },
            "TenantSetActiveRequest": {
                "type": "object",
                "properties": {
                    "tenant_id": {
                        "type": "string",
                        "readOnly": true,
                        "format": "uuid"
                    },
                    "active": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "tenant_id",
                    "active"
                ],
                "example": {
                    "tenant_id": "f7dfed28-ebaa-4bc3-b556-b1787c74e554",
                    "active": true
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Bearer authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "paths": {
        "/tenants/get": {
            "summary": "get tenant information",
            "post": {
                "summary": "POST method",
                "tags": ["tenants"],
                "operationId": "tenant_get_by_id",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantGetInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully retrieved tenant information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "tenant information failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tenants/get/slug": {
            "summary": "get tenant information",
            "post": {
                "summary": "POST method",
                "tags": ["tenants"],
                "operationId": "tenant_get_by_slug",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantGetInfoBySlugRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully retrieved tenant information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "tenant information failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tenants/add": {
            "summary": "create tenant record",
            "post": {
                "summary": "POST method",
                "tags": ["tenants"],
                "operationId": "tenant_add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantAddRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully created tenant record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "add tenant failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tenants/update": {
            "summary": "update tenant record",
            "post": {
                "summary": "POST method",
                "tags": ["tenants"],
                "operationId": "tenant_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully updated tenant record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "add tenant failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tenants/set/active": {
            "summary": "create tenant record",
            "post": {
                "summary": "POST method",
                "tags": ["tenants"],
                "operationId": "tenant_set_active",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantSetActiveRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully set tenant active status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "set active tenant failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/apiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}